+++ spec/javascripts/app/views/stream_post_spec.js
@@ -66,18 +66,16 @@ describe("app.views.StreamPost", function(){
       }});
     });
 
-    context("reshare", function(){
+    context("reshares", function(){
       it("displays a reshare count", function(){
-        this.statusMessage.set({ interactions: {reshares_count : 2 }});
+        this.statusMessage.interactions.set({"reshares_count": 2});
         var view = new this.PostViewClass({model : this.statusMessage}).render();
-
         expect($(view.el).html()).toContain(Diaspora.I18n.t('stream.reshares', {count: 2}));
       });
 
       it("does not display a reshare count for 'zero'", function(){
-        this.statusMessage.interactions.set({ interactions: { reshares_count : 0}} );
+        this.statusMessage.interactions.set({"reshares_count": 0});
         var view = new this.PostViewClass({model : this.statusMessage}).render();
-
         expect($(view.el).html()).not.toContain("0 Reshares");
       });
     });
@@ -86,13 +84,12 @@ describe("app.views.StreamPost", function(){
       it("displays a like count", function(){
         this.statusMessage.interactions.set({likes_count : 1});
         var view = new this.PostViewClass({model : this.statusMessage}).render();
-
         expect($(view.el).html()).toContain(Diaspora.I18n.t('stream.likes', {count: 1}));
       });
+
       it("does not display a like count for 'zero'", function(){
         this.statusMessage.interactions.set({likes_count : 0});
         var view = new this.PostViewClass({model : this.statusMessage}).render();
-
         expect($(view.el).html()).not.toContain("0 Likes");
       });
     });